n_arc02 added to CSharp scripts
n_arc02 script  prompts the user to select three points (start, end, and a point on the arc), calculates the arc’s center and radius, then places subdivided points (and labels) along that arc in the AutoCAD model space. It essentially automates creating a series of points along a specified arc with user-defined division intervals.